Refactor webapp code
Move stuff around
