+++ nbdime/webapp/src/app/common.ts
